requirements.txt
discord.py
sqlalchemy
aiosqlite
colorama
